logist_regression_analysis_for_multiples_of_stocks
Accounting ratios serve as a measure of various economic and financial aspects of companies. Through them, this research aims to identify which aspects are relevant for dividend distribution in Brazil, considering the relevant legal aspects. Through logistic regression analysis, it was found that some variables are statistically significant for this objective: ROE, liquidity, size, investment, revenue instability, and dividend policy stability.

The objective of this work is to replicate the calculated ratios in the article, perform logistic regression analysis for dividend distribution of companies, and also add value to the estimation by calculating another estimated variable for the regression.